requirements.txt
python-telegram-bot==20.7
yt-dlp==2024.4.9
ffmpeg-python==0.2.0
pillow==10.0.0